requirements.txt
aerich==0.5.5
aiofiles==0.7.0
asyncmy==0.1.7
aiomysql==0.0.21
aioredis==2.0.0
aiosmtplib==1.1.6
aiosqlite==0.16.0
appscript==1.1.2
async-timeout==3.0.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.5.30
cffi==1.14.4
click==7.1.2
ddlparse==1.9.0
dnspython==2.0.0
ecdsa==0.17.0
email-validator==1.1.3
fakeredis==1.4.5
fastapi==0.68.0
fastapi-mail==0.4.0
h11==0.11.0
hiredis==2.0.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
img2pdf==0.4.1
iso8601==0.1.13
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
passlib==1.7.4
pikepdf==2.16.1
Pillow==8.3.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
PyMySQL==1.0.2
pyparsing==2.4.7
PyPika==0.44.1
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
python-slugify==5.0.2
pytz==2020.4
redis==3.5.3
rfc3986==1.4.0
rsa==4.7
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.16.0
text-unidecode==1.3
tortoise-orm==0.17.6
tqdm==4.60.0
typing-extensions==3.10.0.0
uvicorn==0.14.0
